+++ Estreya.BlishHUD.Shared/Services/APIStateConfiguration.cs
@@ -1,0 +1,16 @@
+﻿namespace Estreya.BlishHUD.Shared.State;
+
+using Gw2Sharp.WebApi.V2.Models;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
+
+public class APIStateConfiguration : StateConfiguration
+{
+    public List<TokenPermission> NeededPermissions { get; set; } = new List<TokenPermission>();
+
+    public TimeSpan UpdateInterval { get; set; } = Timeout.InfiniteTimeSpan;
+}
